feature: edit and delet own quizes

+++ app/api/quiz/delete/[id]/route.ts
@@ -1,11 +1,27 @@
+import { options } from "@/app/api/auth/[...nextauth]/options";
 import { prisma } from "@/components/prisma";
+import { getServerSession } from "next-auth";
 import { NextResponse } from "next/server";
 
 export async function GET(
-  _request: Request,
-  { params: { id }}: { params: { id: number }}
+  request: Request,
+  { params: { id }}: { params: { id: string }}
 ) {
   try {
+    const session = await getServerSession(options);
+    if (session === null) {
+      throw new Error("There is no session.");
+    }
+
+    if (typeof session.user.name !== "string" && typeof session.user.email !== "string") {
+      throw new Error("Session doesn't have user's data");
+    }
+
+    const quiz = await prisma.quiz.findUnique({ where: { id: +id }});
+
+    if (quiz?.ownerEmail !== session.user.email) {
+      throw new Error(`You cannot remove this quiz. You doesn't own it. Your name is ${session.user.name} but this quiz belong to ${quiz?.ownerName}`);
+    }
     const result = await prisma.quiz.delete({
       where: {
         id: Number(id),
@@ -13,6 +29,6 @@ export async function GET(
     });
     return NextResponse.json(result, { status: 200 });
   } catch (error) {
-    return NextResponse.json({ error }, { status: 500});
+    return NextResponse.json({ error: (error as Error).message  }, { status: 500});
   }
 }
